requirements.txt
Flask==3.1.1
scikit-learn==1.3.2
numpy==1.24.3
requests==2.31.0
gunicorn==20.1.0
python-dotenv==1.0.0
